+++ xbmc/addons/kodi-addon-dev-kit/include/kodi/c-api/CMakeLists.txt
@@ -1,4 +1,5 @@
-set(HEADERS network.h)
+set(HEADERS audio_engine.h
+            network.h)
 
 if(NOT ENABLE_STATIC_LIBS)
   core_add_library(addons_kodi-addon-dev-kit_include_kodi_c-api)
